Rename js folder to modules
